+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-05-18 12:36:08
-// Commit: chore(test): update dependencies and cleanup unused imports in test
+// Updated at: 2026-05-18 13:13:36
+// Commit: fix(test): fix memory leak by cancelling test subscription
 
 class testComponentState {
-  final String commitRef = '1240';
-  final String lastMessage = 'chore(test): update dependencies and cleanup unused imports in test';
-  final DateTime timestamp = DateTime.parse('2026-05-18T12:36:08+05:00');
+  final String commitRef = '1243';
+  final String lastMessage = 'fix(test): fix memory leak by cancelling test subscription';
+  final DateTime timestamp = DateTime.parse('2026-05-18T13:13:36+05:00');
 }